requirements.txt
asn1crypto==1.4.0
astroid==2.7.2
attrs==21.2.0
base45==0.4.2
bidict==0.21.3
cbor2==5.4.1
certifi==2021.5.30
certvalidator==0.11.1
cffi==1.14.6
charset-normalizer==2.0.5
click==8.0.1
cose==0.9.dev8
cryptography==2.8
cycler==0.10.0
Cython==0.29.24
dnspython==2.1.0
ecdsa==0.17.0
eventlet==0.32.0
Flask==2.0.1
Flask-SocketIO==5.1.1
greenlet==1.1.1
idna==3.2
imageio==2.9.0
importlib-metadata==4.8.1
imutils==0.5.4
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib==3.4.3
mccabe==0.6.1
networkx==2.6.2
numpy==1.21.2
opencv-contrib-python==4.1.0.25
oscrypto==1.2.1
picamera==1.13
Pillow==8.3.2
platformdirs==2.2.0
pycparser==2.20
pylint==2.10.2
pyparsing==2.4.7
pytesseract==0.3.8
python-dateutil==2.8.2
python-engineio==4.2.1
python-socketio==5.4.0
PyWavelets==1.1.1
PyYAML==6.0
pyzbar==0.1.8
qrcode==7.3
requests==2.26.0
RPi.GPIO==0.7.0
scikit-image==0.18.3
scipy==1.7.1
screeninfo==0.7
six==1.16.0
spidev==3.5
tesseract==0.1.3
tesseract-ocr==0.0.1
tesserocr==2.5.2
tifffile==2021.8.8
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.1
urllib3==1.26.6
websocket-client==1.2.1
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
